fix(network): cast sndCwnd to uint32 for CongestionWindow field
Type mismatch: sndCwnd is uint8 but CongestionWindow expects uint32.
Discovered by running tests in Docker (golang:1.24 on Linux).

All 75 tests now pass in Docker:
- DNS processor (IPv4/IPv6)
- Link processor (timeouts, states)
- Status processor (HTTP detection)
- Unit tests (helpers, conversions)
- System tests (BTF, kernel, capabilities)
- Performance tests (throughput, memory)
- Negative tests (error handling)

+++ internal/observers/network/observer_ebpf.go
@@ -344,7 +344,7 @@ func (n *NetworkObserver) processRetransmitEvent(ctx context.Context, evt Networ
 					DstPort:          evt.DstPort,
 					RetransmitCount:  uint32(stats.Retransmits),
 					RetransmitRate:   retxRate * 100, // Convert to percentage
-					CongestionWindow: sndCwnd,
+					CongestionWindow: uint32(sndCwnd),
 					ProcessName:      comm,
 				},
 			})
